// src/KRINT.Frontend/src/app/api/model/migrationProgressDto.ts
/**
 * KRINT.API | v1
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ProvisionedDatabaseDto } from './provisionedDatabaseDto';
import { DashboardStatsDtoTotalInstances } from './dashboardStatsDtoTotalInstances';


export interface CleanupStepDto {
    title: string;
    detail: string;
}

export interface MigrationProgressDto {
    step: string;
    status: string;
    message: string;
    currentStep: DashboardStatsDtoTotalInstances;
    totalSteps: DashboardStatsDtoTotalInstances;
    result?: ProvisionedDatabaseDto | null;
    cleanup?: ReadonlyArray<CleanupStepDto> | null;
    error?: string | null;
}